renamed constants directory
